requirements.txt
pymysql
flask
python-dotenv
numpy
opencv-python
matplotlib